feat: Add react-country-flag package for displaying country flags
This commit adds the "react-country-flag" package to the project's dependencies. The package is added with version 3.1.0. This addition enables the application to display country flags based on the selected country in the SignUp page. The "react-country-flag" package provides an easy and efficient way to incorporate country flags into the user interface, enhancing the visual representation of user-selected countries.