feat: Add @radix-ui/react-radio-group package
This commit adds the "@radix-ui/react-radio-group" package to the project's dependencies. The package is added with version 1.1.3. This addition enhances the functionality of the application by providing a radio group component, allowing users to select options in a mutually exclusive manner. The radio group component can be used in various parts of the application, improving user interaction and interface design.